pyproject.toml
[project]
name = "ortools-poc"
version = "0.1.0"
description = "Scheduler backend for vehicle scheduling system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.13",
    "googlemaps==4.10.0",
    "nanoid==2.0.0",
    "ortools==9.9.3963",
    "pydantic==2.11.7",
    "pymongo==4.13.2",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "typing-extensions==4.14.0",
    "uvicorn==0.34.3",
]
